+++ dev-series-tutorials/avail-lens-immersive/package.json
@@ -19,6 +19,9 @@
     "@react-three/drei": "^9.122.0",
     "@react-three/fiber": "^8.18.0",
     "@sparticuz/chromium-min": "^133.0.0",
+    "@tanstack/react-query": "^5.74.4",
+    "@wagmi/core": "^2.16.7",
+    "connectkit": "^1.9.0",
     "ethers": "^6.13.5",
     "framer-motion": "^11.18.2",
     "graphql": "^16.8.1",
@@ -36,6 +39,7 @@
     "three": "^0.162.0",
     "uuid": "^11.1.0",
     "viem": "^2.26.3",
+    "wagmi": "^2.14.16",
     "ws": "^8.13.0",
     "zksync-ethers": "^6.17.0"
   },
